Several big changes related to GP and llamas_snr_full: - Moved gaussian_process from problem/ to approx/ and changed the   include everywhere - Tweaked learn_gp here and there, reduced the plotting, increased the   printing, fixed some inputs - Converted theta to use separate priors for the rgb lenses, which is   much more correct, those lenses are fully different - Changed most of the throughput priors (except collimator, which is   probelmatic) to learn the GP prior from provided throughput files - Changed eta and H accordingly - Fixed the scale in one of the blue lens files
I'm somewhat happy with how this turns out, but I'm a little uncomfortable
fiddling with T_std and the minimum lengthscale of the optimizer.

I'm also very uncomfortable with how often I'm getting unphysically high
throughputs, when that's something in fact to discourage in the interest
of robustness. One potential fix to this is to change the domain of
analysis. If I convert the throughput t[0,1] to a different parameter
u[-inf,inf] with an inverse CDF, for example, then I can train the GP on
function samples drawn from the appropriate domain, and then converting
them back will work fine.

This would involve changing learn_gp to convert t->u, and also would
involve changing gaussian_process to make the right conversions in both
directions. Worried that this is a rabbit hole...